+++ src/Rule/Enum.php
@@ -4,7 +4,7 @@
  * @version   2.0
  * @author    Fuel Development Team
  * @license   MIT License
- * @copyright 2010 - 2016 Fuel Development Team
+ * @copyright 2010 - 2013 Fuel Development Team
  * @link      http://fuelphp.com
  */
 
@@ -12,18 +12,51 @@ namespace Fuel\Validation\Rule;
 
 use Fuel\Validation\AbstractRule;
 
+/**
+ * Checks if a value is one of given values
+ *
+ * @package Fuel\Validation\Rule
+ * @author  Fuel Development Team
+ *
+ * @since   2.0
+ */
 class Enum extends AbstractRule
 {
 
 	/**
-	 * Default failure message
+	 * Contains the rule failure message
 	 *
 	 * @var string
 	 */
-	protected $message = 'The value given is not in the list of allowed values.';
+	protected $message = 'The field is not one of the given value(s).';
 
 	/**
-	 * Sets the values to check against
+	 * Strict check mode
+	 *
+	 * @var bool
+	 */
+	protected $strict = false;
+
+	/**
+	 * Performs validation on the given value
+	 *
+	 * @param mixed  $value Value to validate
+	 * @param string $field Unused by this rule
+	 * @param array  $allFields Unused by this rule
+	 *
+	 * @return bool
+	 *
+	 * @since 2.0
+	 */
+	public function validate($value, $field = null, $allFields = null)
+	{
+		$values = $this->getParameter();
+
+		return in_array($value, $values, $this->strict);
+	}
+
+	/**
+	 * Sets the value(s) and mode to check against
 	 *
 	 * @param string $params
 	 *
@@ -33,35 +66,45 @@ class Enum extends AbstractRule
 	 */
 	public function setParameter($params)
 	{
-		if ( ! is_array($params))
+		if (is_array($params) === false)
 		{
-			$params = [$params];
+			$params = (array) $params;
+		}
+		elseif (array_key_exists('strict', $params) and array_key_exists('values', $params))
+		{
+			$this->strict = (bool) $params['strict'];
+			$params = $params['values'];
 		}
 
 		return parent::setParameter($params);
 	}
 
 	/**
-	 * Returns
+	 * Check strict mode
 	 *
-	 * array(
-	 * 		'values' => <target values>
-	 * );
+	 * @return bool
 	 *
-	 * @return string[]
+	 * @since 2.0
 	 */
-	public function getMessageParameters()
+	public function isStrict()
 	{
-		return array(
-			'values' => implode('|', $this->getParameter()),
-		);
+		return $this->strict;
 	}
 
 	/**
-	 * {@inheritdoc}
+	 * Set strict mode
+	 *
+	 * @param bool $strict
+	 *
+	 * @return $this
+	 *
+	 * @since 2.0
 	 */
-	public function validate($value, $field = null, $allFields = null)
+	public function setStrict($strict)
 	{
-		return in_array($value, $this->getParameter());
+		$this->strict = (bool) $strict;
+
+		return $this;
 	}
+
 }